requirements.txt
Flask==0.12
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
flask-bcrypt